refactor: extract API handlers into api-managers module
Split all 13 service API handlers (Jellyfin, Sonarr, Radarr, qBittorrent,
Proxmox, Portainer, Pi-hole, Glances, Grafana, Nextcloud, Home Assistant,
AdGuard, Immich) out of app.js into individual files under api-managers/,
with a barrel export via api-managers/index.js.

app.js now imports API_HANDLERS from the module, reducing its size
significantly and making each integration independently maintainable.
