+++ api-managers/immich.js
@@ -1,0 +1,13 @@
+export async function api_immich(svc, timedFetch, utils) {
+	try {
+		const b = (svc.endpoint ?? svc.url).replace(/\/$/, '');
+		const headers = svc.api_key ? { 'x-api-key': svc.api_key } : {};
+		const res = await timedFetch(`${b}/api/server-info/stats`, { headers });
+		if (!res.ok) return null;
+		const d = await res.json();
+		return [
+			{ label: 'Photos', value: utils.fmtNum(d.photos) },
+			{ label: 'Videos', value: utils.fmtNum(d.videos) },
+		];
+	} catch { return null; }
+}
